need to refactor
