+++ src/components/news/NewsForm.js
@@ -1,15 +1,73 @@
 import React, { Component } from 'react';
-
+import NewsDataManager from './NewsDataManager'
 import { Button, Modal, ModalHeader, ModalBody, ModalFooter } from 'reactstrap';
 
 
 
 class NewsForm extends Component {
 
+    state = {
+        news: [],
+        title: "",
+        url: "",
+        synopsis: "",
+        timestamp: "",
+        loadingStatus: false,
+    };
+
+    constructor(props) {
+        super(props);
+        this.state = {
+            news: [],
+            title: "",
+            url: "",
+            synopsis: "",
+            timestamp: "",
+            modal: false
+        };
+
+        this.toggle = this.toggle.bind(this);
+    }
+
+    componentDidMount() {
+        NewsDataManager.getAllNews()
+            .then(news => {
+                this.setState({
+                    news: news
+                })
+            })
+    }
+
+    toggle() {
+        this.setState(prevState => ({
+            modal: !prevState.modal
+        }));
+    }
+
+    handleFieldChange = event => {
+        const stateToChange = {};
+        stateToChange[event.target.id] = event.target.value;
+        this.setState(stateToChange);
+    };
 
 
-
-
+    constructNewNewsItem = event => {
+        event.preventDefault();
+        if (this.state.title === ""||
+        this.state.url === "" ||
+        this.state.synopsis === "") {
+            alert("Please fill out all fields");
+        } else {
+            this.setState({ loadingStatus: true });
+            const newsItem = {
+                title: this.state.title,
+                url: this.state.url,
+                synopsis: this.state.synopsis,
+                timestamp: new Date().toLocaleString()
+            };
+            this.addNewsItem
+        }
+    };
 
     render(){
         return(
