+++ src/components/news/NewsList.js
@@ -1,11 +1,21 @@
 import React, { Component } from 'react'
 //import the components we will need
 import NewsCard from './NewsCard'
+import NewsForm from './NewsForm'
 import NewsDataManager from './NewsDataManager'
 import { Button } from 'reactstrap';
 
 
 class NewsList extends Component {
+
+    state = {
+        news: [],
+        title: "",
+        url: "",
+        synopsis: "",
+        timestamp: "",
+        loadingStatus: false,
+    };
 
 
 
@@ -21,21 +31,26 @@ class NewsList extends Component {
 //     })
 //   }
 
+addNewsItem = () => {
 
+}
 
 render(){
     return(
     <React.Fragment>
-      <div className="newsContainerCards">
-        {this.props.state.news.map(newsItem =>
-          <NewsCard
-            key={newsItem.id}
-            news={newsItem}
-            // deleteAnimal={this.deleteAnimal}
-            {...this.props}
-          />
-        )}
-      </div>
+        <NewsForm
+                addNewsItem={this.addNewsItem}
+                {...this.props}/>
+        <div className="newsContainerCards">
+            {this.props.state.news.map(newsItem =>
+            <NewsCard
+                key={newsItem.id}
+                news={newsItem}
+                // deleteAnimal={this.deleteAnimal}
+                {...this.props}
+            />
+            )}
+        </div>
     </React.Fragment>
     )
   }
